refactor: rename topic monitor files and move to own folder
