+++ packages/repocop/src/remediations/topics/topic-monitor-interactive.ts
@@ -2,8 +2,8 @@ import { PublishCommand, SNSClient } from '@aws-sdk/client-sns';
 import type { repocop_github_repository_rules } from '@prisma/client';
 import { awsClientConfig } from 'common/src/aws';
 import { shuffle } from 'common/src/functions';
-import type { Config } from '../config';
-import { removeRepoOwner } from './shared-utilities';
+import type { Config } from '../../config';
+import { removeRepoOwner } from '../shared-utilities';
 
 export function findPotentialInteractives(
 	evaluatedRepos: repocop_github_repository_rules[],
